+++ android/src/main/kotlin/com/example/imagegallerysaver/ImageGallerySaverPlugin.kt
@@ -1,34 +1,38 @@
 package com.example.imagegallerysaver
 
+import android.content.Context
 import android.content.Intent
 import android.content.pm.ApplicationInfo
 import android.content.pm.PackageManager
 import android.graphics.Bitmap
 import android.graphics.BitmapFactory
 import android.net.Uri
-import android.os.Environment
 import android.os.Environment.DIRECTORY_PICTURES
+import android.util.Log
 import androidx.annotation.NonNull
+import io.flutter.embedding.engine.plugins.FlutterPlugin
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
 import io.flutter.plugin.common.MethodChannel.Result
 import io.flutter.plugin.common.MethodCall
-import io.flutter.plugin.common.PluginRegistry.Registrar
 import java.io.File
 import java.io.FileOutputStream
 import java.io.IOException
 
-class ImageGallerySaverPlugin(private val registrar: Registrar) : MethodCallHandler {
+class ImageGallerySaverPlugin : FlutterPlugin, MethodCallHandler {
 
-    companion object {
-        @JvmStatic
-        fun registerWith(registrar: Registrar) {
-            val channel = MethodChannel(registrar.messenger(), "image_gallery_saver")
-            channel.setMethodCallHandler(ImageGallerySaverPlugin(registrar))
-        }
+    private var context: Context? = null
+    private lateinit var channel: MethodChannel
+
+    override fun onAttachedToEngine(@NonNull flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
+        Log.d("ImageGallerySaverPlugin", "onAttachedToEngine")
+        context = flutterPluginBinding.applicationContext
+        channel = MethodChannel(flutterPluginBinding.binaryMessenger, "image_gallery_saver")
+        channel.setMethodCallHandler(this)
     }
 
     override fun onMethodCall(@NonNull call: MethodCall, @NonNull result: Result) {
+        Log.d("ImageGallerySaverPlugin", "onMethodCall, method = ${call.method}")
         when (call.method) {
             "saveImageToGallery" -> {
                 val image = call.argument<ByteArray>("imageBytes") ?: return
@@ -44,21 +48,22 @@ class ImageGallerySaverPlugin(private val registrar: Registrar) : MethodCallHand
         }
     }
 
+    override fun onDetachedFromEngine(@NonNull binding: FlutterPlugin.FlutterPluginBinding) {
+        context = null
+        channel.setMethodCallHandler(null)
+    }
+
     private fun generateFile(extension: String = "", name: String? = null): File? {
-        val pictureDir = Environment.getExternalStoragePublicDirectory(DIRECTORY_PICTURES)
+        val pictureDir = context?.getExternalFilesDir(DIRECTORY_PICTURES)
         pictureDir?.run {
             if (!pictureDir.exists()) {
                 pictureDir.mkdirs()
-            }
-            val appDir = File(pictureDir, getApplicationName())
-            if (!appDir.exists()) {
-                appDir.mkdir()
             }
             var fileName = name ?: System.currentTimeMillis().toString()
             if (extension.isNotEmpty()) {
                 fileName += (".$extension")
             }
-            val file = File(appDir, fileName)
+            val file = File(pictureDir, fileName)
             if (!file.exists()) {
                 file.createNewFile()
             }
@@ -68,7 +73,6 @@ class ImageGallerySaverPlugin(private val registrar: Registrar) : MethodCallHand
     }
 
     private fun saveImageToGallery(bmp: Bitmap, quality: Int, name: String?): String {
-        val context = registrar.activeContext().applicationContext
         val file = generateFile("jpg", name = name)
         try {
             val fos = FileOutputStream(file)
@@ -77,7 +81,7 @@ class ImageGallerySaverPlugin(private val registrar: Registrar) : MethodCallHand
             fos.flush()
             fos.close()
             val uri = Uri.fromFile(file)
-            context.sendBroadcast(Intent(Intent.ACTION_MEDIA_SCANNER_SCAN_FILE, uri))
+            context?.sendBroadcast(Intent(Intent.ACTION_MEDIA_SCANNER_SCAN_FILE, uri))
             bmp.recycle()
             return uri.toString()
         } catch (e: IOException) {
@@ -87,14 +91,13 @@ class ImageGallerySaverPlugin(private val registrar: Registrar) : MethodCallHand
     }
 
     private fun saveFileToGallery(filePath: String): String {
-        val context = registrar.activeContext().applicationContext
         return try {
             val originalFile = File(filePath)
             val file = generateFile(originalFile.extension)
             if (file != null) {
                 originalFile.copyTo(file)
                 val uri = Uri.fromFile(file)
-                context.sendBroadcast(Intent(Intent.ACTION_MEDIA_SCANNER_SCAN_FILE, uri))
+                context?.sendBroadcast(Intent(Intent.ACTION_MEDIA_SCANNER_SCAN_FILE, uri))
                 uri.toString()
             }
             ""
@@ -105,15 +108,13 @@ class ImageGallerySaverPlugin(private val registrar: Registrar) : MethodCallHand
     }
 
     private fun getApplicationName(): String {
-        val context = registrar.activeContext().applicationContext
         var ai: ApplicationInfo? = null
         try {
-            ai = context.packageManager.getApplicationInfo(context.packageName, 0)
+            ai = context?.packageManager?.getApplicationInfo(context?.packageName, 0)
         } catch (e: PackageManager.NameNotFoundException) {
         }
         return if (ai != null) {
-            val charSequence = context.packageManager.getApplicationLabel(ai)
-            StringBuilder(charSequence.length).append(charSequence).toString()
+            context?.packageManager?.getApplicationLabel(ai)?.toString() ?: "image_gallery_saver"
         } else {
             "image_gallery_saver"
         }